pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Source code for the mini-paper 'Beyond Pruning and Dropout: Evolving Robust Networks via Persistent Stochastic Ablation'"
authors = ["Tim Cotten <tcotten@scrypted.ai>", "Tim Cotten <tcotten2@gmu.edu>"]
readme = "README.md"
packages = [{include = "src"}, {include = "aws"}]



[tool.poetry.dependencies]
python = ">=3.9,<3.12"
torch = "^2.7.1"
torchvision = "^0.22.1"
rich = "^14.1.0"
safetensors = "^0.5.3"
python-dotenv = "^1.1.1"
boto3 = "^1.40.1"
botocore = "^1.40.1"
sagemaker = "^2.249.0"
matplotlib = "^3.8.0"
scipy = "1.10.1"
adjusttext = "^1.3.0"
shapely = "<2.0"

[tool.poetry.scripts]
train = "src.train_psa_simplemlp:main"
clean = "scripts.clean:main"
clean-all = "scripts.clean_all:main"
generate-reproduction-tests = "scripts.generate_reproduction_tests:main"
sagemaker-results-parser = "aws.sagemaker.results_parser:main"
sagemaker-estimate-storage = "aws.sagemaker.storage_estimator:main"
sagemaker-get-training-logs = "aws.sagemaker.log_downloader:main"
make-architecture-table = "scripts.make_table_architectures:main"
make-trial-accuracy-table = "scripts.make_table_trial_accuracy:main"
make-figure-design-space = "scripts.make_figure_design_space:main"
make-figure-heatmaps = "scripts.make_figure_heatmaps:main"
analyze-dataset = "analyze_dataset:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
